SBTMM2-71  Update TankNotifierApi app.
Updated the SkyBitzApiController ValidateTankId method:  The cache uses a new prefix key - so as not to collide with older cache etries and uses a new cache object - TankUtilityCacheDevice - to store cache information.

The cache device encapsulates the long and short Tank Utility Id's and an "Is Valid" property.

Also minor updates to the ServiceBusTopicSender class;  No major changes just some code re-org and the setting of the ContentType in the SendMessage method.