Update app.configUpdate web.config
Added the sbtanksoftwareprod@ametek.com and sbtanksoftwaredev@ametek.com users to the web.config.

The PROD deployment should use the sbtanksoftwareprod@ametek.com account since the devices on that account are actually available.

The PVT deployment should use the sbtanksoftwareprod@ametek.com account since the devices on that account are actually available.